// src/app/models/tag.model.ts
export class tagsmodel {
  Id: number;
  Name: string;
  IsActive: boolean;
  Description: string;
  CreatedById: number;
  UpdatedById: number;
  CreatedByRoleId: number;
  UpdatedByRoleId: number;
  ParameterCarrier: string;
  ProjectId: number;
  LoggedInScopeEntityType: string;
  LoggedInScopeEntityId: number;
  TagCategoryId: number;
}